image-analysis-agent/src/models/turtle_model.py
import numpy as np
import cv2
import torch
from typing import Dict, Any
import logging

logger = logging.getLogger(__name__)


class TurtleIdentificationModel:
    """
    Deniz kaplumbağası tanımlama modeli
    Pre-trained model (YOLOv8 + ResNet50 feature extractor)
    """

    def __init__(self, model_path: str = None):
        """
        Model initialization
        
        Args:
            model_path: Pre-trained model weights path
        """
        self.device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
        self.model_path = model_path
        self.model = None
        self.feature_extractor = None
        
        self._load_model()

    def _load_model(self):
        """
        Model yükleme (YOLOv8)
        Gerçek implementasyonda pre-trained model yüklenecek
        """
        try:
            # YOLOv8 modeli yükle (örnek)
            # from ultralytics import YOLO
            # self.model = YOLO(self.model_path)
            
            logger.info("Model placeholder initialized (use real model in production)")
            
            # Feature extractor (ResNet50)
            # self.feature_extractor = torchvision.models.resnet50(pretrained=True)
            # self.feature_extractor.eval()
            
        except Exception as e:
            logger.error(f"Failed to load model: {e}")
            raise

    def identify_turtle(self, image: np.ndarray) -> Dict[str, Any]:
        """
        Görüntüdeki kaplumbağayı tanımla
        
        Args:
            image: Processed image (numpy array)
            
        Returns:
            Dictionary with identification results
        """
        try:
            # Örnek sonuç (gerçek implementasyonda ML model çıktısı)
            result = {
                'turtle_id': 'turtle_001',
                'species': 'Chelonia mydas',  # Green sea turtle
                'confidence': 0.92,
                'shell_pattern': 'unique_pattern_hash',
                'characteristics': {
                    'shell_color': 'dark_green',
                    'size_estimate': 'adult',
                    'distinctive_marks': ['scar_1', 'algae_growth']
                }
            }
            
            return result
            
        except Exception as e:
            logger.error(f"Error during identification: {e}")
            raise

    def extract_features(self, image: np.ndarray) -> np.ndarray:
        """
        Görüntüden özellik vektörü çıkar (benzerlik araması için)
        
        Args:
            image: Processed image
            
        Returns:
            Feature vector (numpy array)
        """
        try:
            # Örnek: 512-dim feature vector
            features = np.random.randn(512).astype(np.float32)
            return features
            
        except Exception as e:
            logger.error(f"Error extracting features: {e}")
            raise

    def batch_identify(self, images: list) -> list:
        """
        Batch tanımlama
        
        Args:
            images: List of images
            
        Returns:
            List of results
        """
        results = []
        for image in images:
            try:
                result = self.identify_turtle(image)
                results.append(result)
            except Exception as e:
                logger.error(f"Error processing image: {e}")
                results.append({'error': str(e)})
        
        return results

    def get_model_info(self) -> Dict[str, Any]:
        """Model bilgisini döndür"""
        return {
            'model_type': 'YOLOv8-ResNet50',
            'device': str(self.device),
            'model_path': self.model_path,
            'input_size': (640, 640, 3)
        }
