image-analysis-agent/requirements.txt
flask==2.3.2
flask-cors==4.0.0
opencv-python==4.7.0.72
numpy==1.24.3
tensorflow==2.13.0
torch==2.0.1
torchvision==0.15.2
python-dotenv==1.0.0
requests==2.31.0
pillow==10.0.0
scipy==1.11.1
scikit-image==0.21.0
